Move `CollateralInputs` to `Cardano.Wallet.Read.Tx.CollateralInputs`
