+++ lib/primitive/lib/Cardano/Wallet/Primitive/Ledger/Read/Tx.hs
@@ -16,6 +16,9 @@ import Prelude
 import Cardano.Read.Ledger.Tx.CBOR
     ( renderTxToCBOR
     )
+import Cardano.Read.Ledger.Tx.CollateralInputs
+    ( getEraCollateralInputs
+    )
 import Cardano.Read.Ledger.Tx.Inputs
     ( getEraInputs
     )
@@ -52,9 +55,6 @@ import Cardano.Wallet.Read
     )
 import Cardano.Wallet.Read.Eras
     ( eraValue
-    )
-import Cardano.Wallet.Read.Tx.CollateralInputs
-    ( getEraCollateralInputs
     )
 import Cardano.Wallet.Read.Tx.CollateralOutputs
     ( getEraCollateralOutputs
