+++ lib/wallet/src/Cardano/Wallet/DB/Store/Transactions/TransactionInfo.hs
@@ -18,6 +18,9 @@ import Cardano.Read.Ledger.Tx.CBOR
     ( TxCBOR
     , renderTxToCBOR
     )
+import Cardano.Read.Ledger.Tx.CollateralInputs
+    ( getEraCollateralInputs
+    )
 import Cardano.Read.Ledger.Tx.Inputs
     ( getEraInputs
     )
@@ -94,9 +97,6 @@ import Cardano.Wallet.Read.Eras
     ( EraValue
     , IsEra
     , applyEraFun
-    )
-import Cardano.Wallet.Read.Tx.CollateralInputs
-    ( getEraCollateralInputs
     )
 import Cardano.Wallet.Read.Tx.CollateralOutputs
     ( getEraCollateralOutputs
